requirements.txt
rpi-rfm69
paho-mqtt